requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-text-splitters
faiss-cpu
pandas
python-dotenv
pypdf